chore: simplify codes, fix directory
